WIP: - move detection to different import - throw better errors from prerender - dryrun no longer prints output
