+++ packages/prerender/detection/index.js
@@ -1,0 +1,2 @@
+/* eslint-env es6, commonjs */
+module.exports = require('../dist/detection')
