+++ packages/prerender/detection/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "./index.js"
+}
