+++ packages/prerender/src/index.ts
@@ -1,2 +1,1 @@
 export * from './runPrerender'
-export * from './router'
